[Docs]readme and requirements modify
Co-authored-by: peng-hengduo<penghengduo@huawei.com>



# message auto-generated for no-merge-commit merge:
!1578 merge qwenimage_update into master

[Docs]readme and requirements modify

Created-by: peng-hengduo
Commit-by: peng-hengduo
Merged-by: ascend-robot
Description: ## Motivation

The statements in the README have been refined to make them clearer, corrected the transformers version in the requirements.

## Modification

Modified the readme.md for qwen image, corrected the transformers version in the requirements.

## Self-test (Optional)

If modifications to this PR may cause/fix function/accuracy/performance DTSs/issues, a self-inspection record needs to be attached.

## BC-breaking (Optional)

If there are compatibility issues, such as dependencies on cann/torch_npu versions, they need to be explained in the PR.

## Checklist

**Before PR**:

- [ ] The new code needs to comply with the Clean Code specification.
- [ ] The PR content is self-checked, and the expression can be clear and the writing standardized

**After PR**:

- [ ] CLA has been signed and all committers have signed the CLA in this PR.
- [ ] The ci-pipeline is passed, Code Check	is passed.


See merge request: Ascend/MindSpeed-MM!1578