【2.2.0】【Docs】【mindspore】update doc
Co-authored-by: AACAES<liuyan7892@163.com>



# message auto-generated for no-merge-commit merge:
!1607 merge 2.2.0 into 2.2.0

【2.2.0】【Docs】【mindspore】update doc

Created-by: AACAES
Commit-by: AACAES
Merged-by: ascend-robot
Description: ## Motivation

Update documentation to align with the current state of the repository.

## Modification

Update dependency versions and refine documentation wording.

## Self-test (Optional)

If modifications to this PR may cause/fix function/accuracy/performance DTSs/issues, a self-inspection record needs to be attached.

## BC-breaking (Optional)

If there are compatibility issues, such as dependencies on cann/torch_npu versions, they need to be explained in the PR.

## Checklist

**Before PR**:

- [x] The new code needs to comply with the Clean Code specification.
- [x] The PR content is self-checked, and the expression can be clear and the writing standardized

**After PR**:

- [ ] CLA has been signed and all committers have signed the CLA in this PR.
- [ ] The ci-pipeline is passed, Code Check	is passed.


See merge request: Ascend/MindSpeed-MM!1607