[bugfix]Modify the description of the default loss calculation behavior for vlm_model
Co-authored-by: MoCuishle-M<1656631289@qq.com>



# message auto-generated for no-merge-commit merge:
!1609 merge 2.2.0 into 2.2.0

[bugfix]Modify the description of the default loss calculation behavior for vlm_model

Created-by: MoCuishle-M
Commit-by: MoCuishle-M
Merged-by: ascend-robot
Description: ## Motivation

Modify the description of the default loss calculation behavior for vlm_model.

## Modification

Modify the description of the default loss calculation behavior for vlm_model.

## Self-test (Optional)

If modifications to this PR may cause/fix function/accuracy/performance DTSs/issues, a self-inspection record needs to be attached.

## BC-breaking (Optional)

If there are compatibility issues, such as dependencies on cann/torch_npu versions, they need to be explained in the PR.

## Checklist

**Before PR**:

- [ ] The new code needs to comply with the Clean Code specification.
- [ ] The PR content is self-checked, and the expression can be clear and the writing standardized

**After PR**:

- [ ] CLA has been signed and all committers have signed the CLA in this PR.
- [ ] The ci-pipeline is passed, Code Check	is passed.


See merge request: Ascend/MindSpeed-MM!1609