[bugfix]Use getattr to get arguments in the validate_args function.
Co-authored-by: MoCuishle-M<1656631289@qq.com>



# message auto-generated for no-merge-commit merge:
!1615 merge 2.2.0 into 2.2.0

[bugfix]Use getattr to get arguments in the validate_args function.

Created-by: MoCuishle-M
Commit-by: MoCuishle-M
Merged-by: ascend-robot
Description: ## Motivation

Use getattr to get arguments in the validate_args function. 
In some pipelines, such as feature extraction, custom arguments are not added to the namespace; using getattr makes the code more robust.

## Modification

Use getattr to get arguments in the validate_args function. 

## Self-test (Optional)

If modifications to this PR may cause/fix function/accuracy/performance DTSs/issues, a self-inspection record needs to be attached.

## BC-breaking (Optional)

If there are compatibility issues, such as dependencies on cann/torch_npu versions, they need to be explained in the PR.

## Checklist

**Before PR**:

- [x] The new code needs to comply with the Clean Code specification.
- [x] The PR content is self-checked, and the expression can be clear and the writing standardized

**After PR**:

- [x] CLA has been signed and all committers have signed the CLA in this PR.
- [x] The ci-pipeline is passed, Code Check	is passed.


See merge request: Ascend/MindSpeed-MM!1615