[Docs] Supplement the missing information for 2.2.0 and fix support for version of Python
Co-authored-by: AZe_404<wangze62@h-partners.com>



# message auto-generated for no-merge-commit merge:
!1686 merge 220info into master

[Docs] Supplement the missing information for 2.2.0 and fix support for version of Python

Created-by: AZe_404
Commit-by: AZe_404
Merged-by: ascend-robot
Description: ## Motivation

Supplement the missing information for 2.2.0 and fix support for version of Python.

## Modification

Supplement the missing information for 2.2.0 and fix support for version of Python.

## Self-test (Optional)

If modifications to this PR may cause/fix function/accuracy/performance DTSs/issues, a self-inspection record needs to be attached.

## BC-breaking (Optional)

If there are compatibility issues, such as dependencies on cann/torch_npu versions, they need to be explained in the PR.

## Checklist

**Before PR**:

- [x] The new code needs to comply with the Clean Code specification.
- [x] The PR content is self-checked, and the expression can be clear and the writing standardized

**After PR**:

- [x] CLA has been signed and all committers have signed the CLA in this PR.
- [x] The ci-pipeline is passed, Code Check	is passed.


See merge request: Ascend/MindSpeed-MM!1686