[Bugfix] Fix the internvl3.5 code, which affects qwen3vl
Co-authored-by: wumengjie (D)<wumengjie6@huawei.com>



# message auto-generated for no-merge-commit merge:
!1746 merge master into master

[Bugfix] Fix the internvl3.5 code, which affects qwen3vl

Created-by: meng-coding
Commit-by: wumengjie (D)
Merged-by: ascend-robot
Description: ## Motivation

Fix the internvl3.5 code, which affects qwen3vl

## Modification

Modify the import of the InternVLChatModel class in internvl3.5.

## Self-test (Optional)

If modifications to this PR may cause/fix function/accuracy/performance DTSs/issues, a self-inspection record needs to be attached.

## BC-breaking (Optional)

If there are compatibility issues, such as dependencies on cann/torch_npu versions, they need to be explained in the PR.

## Checklist

**Before PR**:

- [ ] The new code needs to comply with the Clean Code specification.
- [ ] The PR content is self-checked, and the expression can be clear and the writing standardized

**After PR**:

- [ ] CLA has been signed and all committers have signed the CLA in this PR.
- [ ] The ci-pipeline is passed, Code Check	is passed.


See merge request: Ascend/MindSpeed-MM!1746