[Feature]Added Qwen Image Edit training and inference
Co-authored-by: peng-hengduo<penghengduo@huawei.com>



# message auto-generated for no-merge-commit merge:
!1773 merge add_qwen_image_edit into master

[Feature]Added Qwen Image Edit training and inference

Created-by: peng-hengduo
Commit-by: peng-hengduo
Merged-by: ascend-robot
Description: ## Motivation

Merge the training and inference of the Qwen image edit model, based on the original Diffsynth repository.

## Modification

Merge the training and inference of the Qwen image edit model, based on the original Diffsynth repository.

## Self-test (Optional)

If modifications to this PR may cause/fix function/accuracy/performance DTSs/issues, a self-inspection record needs to be attached.

## BC-breaking (Optional)

If there are compatibility issues, such as dependencies on cann/torch_npu versions, they need to be explained in the PR.

## Checklist

**Before PR**:

- [x] The new code needs to comply with the Clean Code specification.
- [x] The PR content is self-checked, and the expression can be clear and the writing standardized

**After PR**:

- [x] CLA has been signed and all committers have signed the CLA in this PR.
- [x] The ci-pipeline is passed, Code Check	is passed.


See merge request: Ascend/MindSpeed-MM!1773