[master][Feature]Support swap attention for Qwen2.5vl 72B
Co-authored-by: iquoyuw<wuyouqi1@h-partners.com>



# message auto-generated for no-merge-commit merge:
!1775 merge sa-master into master

[mindspore][Feature]Support swap attention for Qwen2.5vl 72B

Created-by: weixin_47897441
Commit-by: iquoyuw
Merged-by: ascend-robot
Description: ## Motivation

Add support for the swap attention feature to the Qwen2.5vl 72B model, enabling compatibility with the MindSpore framework.

## Modification

This PR introduces a new** --swap-attention** configuration option, which aligns the MindSpore-based finetuning script with PTA.

## Self-test (Optional)

If modifications to this PR may cause/fix function/accuracy/performance DTSs/issues, a self-inspection record needs to be attached.

## BC-breaking (Optional)

If there are compatibility issues, such as dependencies on cann/torch_npu versions, they need to be explained in the PR.

## Checklist

**Before PR**:

- [ ] The new code needs to comply with the Clean Code specification.
- [ ] The PR content is self-checked, and the expression can be clear and the writing standardized

**After PR**:

- [ ] CLA has been signed and all committers have signed the CLA in this PR.
- [ ] The ci-pipeline is passed, Code Check	is passed.


See merge request: Ascend/MindSpeed-MM!1775