[Docs]modify qwen image edit in readme
Co-authored-by: peng-hengduo<penghengduo@huawei.com>



# message auto-generated for no-merge-commit merge:
!1798 merge modify_qwenimageedit into master

[Docs]modify qwen image edit in readme

Created-by: peng-hengduo
Commit-by: peng-hengduo
Merged-by: ascend-robot
Description: ## Motivation

Wrote the wrong name for the Qwen image edit model; here is the correction.

## Modification
Wrote the wrong name for the Qwen image edit model; here is the correction.

## Self-test (Optional)

If modifications to this PR may cause/fix function/accuracy/performance DTSs/issues, a self-inspection record needs to be attached.

## BC-breaking (Optional)

If there are compatibility issues, such as dependencies on cann/torch_npu versions, they need to be explained in the PR.

## Checklist

**Before PR**:

- [ ] The new code needs to comply with the Clean Code specification.
- [ ] The PR content is self-checked, and the expression can be clear and the writing standardized

**After PR**:

- [ ] CLA has been signed and all committers have signed the CLA in this PR.
- [ ] The ci-pipeline is passed, Code Check	is passed.


See merge request: Ascend/MindSpeed-MM!1798