[master][Mindspore][Modify]Remove recompute settings
Co-authored-by: iquoyuw<wuyouqi1@h-partners.com>



# message auto-generated for no-merge-commit merge:
!1850 merge swap-attention into master

[master][Mindspore][Modify]Remove recompute settings

Created-by: weixin_47897441
Commit-by: iquoyuw
Merged-by: ascend-robot
Description: ## Motivation

The motivation of this PR is to resolve the configuration conflict between the newly added and enabled-by-default swap attention feature in the SFT script and the full recompute settings. To align with the model configuration in PTA and eliminate the conflict, we need to remove the recompute-related configurations.

## Modification

The key modification in this PR is removing the recompute settings from the model.json file. This change ensures compatibility with the default-enabled swap attention feature in the SFT script (as the two configurations are mutually exclusive) and keeps the model configuration consistent with PTA's model.json.

## Self-test (Optional)

After removing the recompute settings, the SFT script can successfully launch and run the model without any functional exceptions or compatibility issues.


## BC-breaking (Optional)

If there are compatibility issues, such as dependencies on cann/torch_npu versions, they need to be explained in the PR.

## Checklist

**Before PR**:

- [ ] The new code needs to comply with the Clean Code specification.
- [ ] The PR content is self-checked, and the expression can be clear and the writing standardized

**After PR**:

- [ ] CLA has been signed and all committers have signed the CLA in this PR.
- [ ] The ci-pipeline is passed, Code Check	is passed.


See merge request: Ascend/MindSpeed-MM!1850