[Modify] Specify the value of sharding_size for wan2.2
Co-authored-by: AZe_404<wangze62@h-partners.com>



# message auto-generated for no-merge-commit merge:
!1876 merge fsdp2_sharding_size_230 into 2.3.0

[Modify] Specify the value of sharding_size for wan2.2

Created-by: AZe_404
Commit-by: AZe_404
Merged-by: ascend-robot
Description: ## Motivation

Specify the value of sharding_size for wan2.2.

## Modification

Specify the value of sharding_size for wan2.2.

## Self-test (Optional)

If modifications to this PR may cause/fix function/accuracy/performance DTSs/issues, a self-inspection record needs to be attached.

## BC-breaking (Optional)

If there are compatibility issues, such as dependencies on cann/torch_npu versions, they need to be explained in the PR.

## Checklist

**Before PR**:

- [x] The new code needs to comply with the Clean Code specification.
- [x] The PR content is self-checked, and the expression can be clear and the writing standardized

**After PR**:

- [x] CLA has been signed and all committers have signed the CLA in this PR.
- [x] The ci-pipeline is passed, Code Check	is passed.


See merge request: Ascend/MindSpeed-MM!1876