[feature] qwen3vl add synchronze
Co-authored-by: cxiaolong<2845907121@qq.com>



# message auto-generated for no-merge-commit merge:
!1892 merge master into master

[feature] qwen3vl add synchronze

Created-by: cxiaolong
Commit-by: cxiaolong
Merged-by: ascend-robot
Description: ## Motivation

Please describe the motivation of this PR and the goal you want to achieve through this PR.

## Modification

add synchronze per layer in vit and llm, for memory reuse between streams when using FSDP2.

## Self-test (Optional)

If modifications to this PR may cause/fix function/accuracy/performance DTSs/issues, a self-inspection record needs to be attached.

## BC-breaking (Optional)

If there are compatibility issues, such as dependencies on cann/torch_npu versions, they need to be explained in the PR.

## Checklist

**Before PR**:

- [x] The new code needs to comply with the Clean Code specification.
- [x] The PR content is self-checked, and the expression can be clear and the writing standardized

**After PR**:

- [x] CLA has been signed and all committers have signed the CLA in this PR.
- [x] The ci-pipeline is passed, Code Check	is passed.


See merge request: Ascend/MindSpeed-MM!1892