[Docs] Update the version of CANN in installation part
Co-authored-by: AZe_404<wangze62@h-partners.com>



# message auto-generated for no-merge-commit merge:
!1953 merge update_installation into master

[Docs] Update the version of CANN in installation part

Created-by: AZe_404
Commit-by: AZe_404
Merged-by: ascend-robot
Description: ## Motivation

Update the version of CANN in installation part.

## Modification

Update the version of CANN in installation part.

## Self-test (Optional)

If modifications to this PR may cause/fix function/accuracy/performance DTSs/issues, a self-inspection record needs to be attached.

## BC-breaking (Optional)

If there are compatibility issues, such as dependencies on cann/torch_npu versions, they need to be explained in the PR.

## Checklist

**Before PR**:

- [x] The new code needs to comply with the Clean Code specification.
- [x] The PR content is self-checked, and the expression can be clear and the writing standardized

**After PR**:

- [x] CLA has been signed and all committers have signed the CLA in this PR.
- [x] The ci-pipeline is passed, Code Check	is passed.


See merge request: Ascend/MindSpeed-MM!1953